# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NearConnect API - local events discovery and community platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface LoginInput {
  /** Email or phone number */
  identifier: string;
  password: string;
}

export interface RegisterInput {
  name: string;
  /** Email or phone number */
  identifier: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  attributes?: string[];
  eventsHosted?: number;
  eventsJoined?: number;
  createdAt?: string;
}

export interface AuthResult {
  user: User;
  isNewUser: boolean;
}

export interface UserUpdate {
  name?: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  attributes?: string[];
}

export type EventType = typeof EventType[keyof typeof EventType];


export const EventType = {
  morning_walk: 'morning_walk',
  gym: 'gym',
  tournament: 'tournament',
  trip: 'trip',
  picnic: 'picnic',
  social: 'social',
  sos: 'sos',
  other: 'other',
} as const;

export type EventStatus = typeof EventStatus[keyof typeof EventStatus];


export const EventStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  type: EventType;
  hostId: number;
  host?: User;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  startAt: string;
  /** @nullable */
  endAt?: string | null;
  /** @nullable */
  maxParticipants?: number | null;
  currentParticipants?: number;
  /** @nullable */
  entryFee?: number | null;
  status: EventStatus;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  isTournament?: boolean;
  /** @nullable */
  prize?: string | null;
  createdAt?: string;
}

export type EventInputType = typeof EventInputType[keyof typeof EventInputType];


export const EventInputType = {
  morning_walk: 'morning_walk',
  gym: 'gym',
  tournament: 'tournament',
  trip: 'trip',
  picnic: 'picnic',
  social: 'social',
  sos: 'sos',
  other: 'other',
} as const;

export interface EventInput {
  title: string;
  /** @nullable */
  description?: string | null;
  type: EventInputType;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  startAt: string;
  /** @nullable */
  endAt?: string | null;
  /** @nullable */
  maxParticipants?: number | null;
  /** @nullable */
  entryFee?: number | null;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  isTournament?: boolean;
  /** @nullable */
  prize?: string | null;
}

export type EventUpdateStatus = typeof EventUpdateStatus[keyof typeof EventUpdateStatus];


export const EventUpdateStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface EventUpdate {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  location?: string | null;
  startAt?: string;
  /** @nullable */
  endAt?: string | null;
  /** @nullable */
  maxParticipants?: number | null;
  /** @nullable */
  entryFee?: number | null;
  status?: EventUpdateStatus;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  prize?: string | null;
}

export interface EventTypeCount {
  type: string;
  count: number;
}

export interface EventStats {
  totalEvents: number;
  upcomingEvents: number;
  totalUsers: number;
  activeToday: number;
  byType?: EventTypeCount[];
}

export type RegistrationStatus = typeof RegistrationStatus[keyof typeof RegistrationStatus];


export const RegistrationStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Registration {
  id: number;
  eventId: number;
  userId: number;
  user?: User;
  event?: Event;
  status: RegistrationStatus;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface RegistrationInput {
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type ChatRoomType = typeof ChatRoomType[keyof typeof ChatRoomType];


export const ChatRoomType = {
  event: 'event',
  direct: 'direct',
  group: 'group',
} as const;

export interface ChatRoom {
  id: number;
  name: string;
  type: ChatRoomType;
  /** @nullable */
  eventId?: number | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  lastMessage?: string | null;
  lastMessageAt: string;
  unreadCount?: number;
  participantCount?: number;
}

export interface ChatMessage {
  id: number;
  roomId: number;
  userId: number;
  user?: User;
  text: string;
  createdAt: string;
}

export interface ChatMessageInput {
  text: string;
}

export type NoticeCategory = typeof NoticeCategory[keyof typeof NoticeCategory];


export const NoticeCategory = {
  general: 'general',
  lost_found: 'lost_found',
  help_wanted: 'help_wanted',
  announcement: 'announcement',
  urgent: 'urgent',
} as const;

export interface Notice {
  id: number;
  title: string;
  content: string;
  authorId: number;
  author?: User;
  category: NoticeCategory;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export type NoticeInputCategory = typeof NoticeInputCategory[keyof typeof NoticeInputCategory];


export const NoticeInputCategory = {
  general: 'general',
  lost_found: 'lost_found',
  help_wanted: 'help_wanted',
  announcement: 'announcement',
  urgent: 'urgent',
} as const;

export interface NoticeInput {
  title: string;
  content: string;
  category: NoticeInputCategory;
  /** @nullable */
  imageUrl?: string | null;
}

export type SosRequestStatus = typeof SosRequestStatus[keyof typeof SosRequestStatus];


export const SosRequestStatus = {
  open: 'open',
  resolved: 'resolved',
  cancelled: 'cancelled',
} as const;

export interface SosRequest {
  id: number;
  title: string;
  description: string;
  requesterId: number;
  requester?: User;
  /** @nullable */
  location?: string | null;
  status: SosRequestStatus;
  responseCount?: number;
  createdAt: string;
}

export interface SosRequestInput {
  title: string;
  description: string;
  /** @nullable */
  location?: string | null;
}

export interface SosResponseInput {
  message: string;
}

export type ListEventsParams = {
type?: ListEventsType;
lat?: number;
lng?: number;
radius?: number;
limit?: number;
offset?: number;
};

export type ListEventsType = typeof ListEventsType[keyof typeof ListEventsType];


export const ListEventsType = {
  morning_walk: 'morning_walk',
  gym: 'gym',
  tournament: 'tournament',
  trip: 'trip',
  picnic: 'picnic',
  social: 'social',
  sos: 'sos',
  other: 'other',
} as const;

export type GetFeedParams = {
limit?: number;
};

export type ListNoticesParams = {
category?: ListNoticesCategory;
limit?: number;
};

export type ListNoticesCategory = typeof ListNoticesCategory[keyof typeof ListNoticesCategory];


export const ListNoticesCategory = {
  general: 'general',
  lost_found: 'lost_found',
  help_wanted: 'help_wanted',
  announcement: 'announcement',
  urgent: 'urgent',
} as const;