# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NearConnect API - local events discovery and community platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user's public profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /events/{eventId}/registrations:
    get:
      operationId: getEventRegistrations
      tags: [registrations]
      summary: List registrations for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
    post:
      operationId: registerForEvent
      tags: [registrations]
      summary: Register for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Already registered or full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{eventId}/registrations/me:
    delete:
      operationId: cancelRegistration
      tags: [registrations]
      summary: Cancel my registration for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    User:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        attributes:
          type: array
          items:
            type: string
        eventsHosted:
          type: integer
        eventsJoined:
          type: integer
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        attributes:
          type: array
          items:
            type: string
    Event:
      type: object
      required: [id, title, type, hostId, status, startAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [morning_walk, gym, tournament, trip, picnic, social, sos, other]
        hostId:
          type: integer
        host:
          $ref: "#/components/schemas/User"
        location:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        startAt:
          type: string
        endAt:
          type: ["string", "null"]
        maxParticipants:
          type: ["integer", "null"]
        currentParticipants:
          type: integer
        entryFee:
          type: ["number", "null"]
        status:
          type: string
          enum: [upcoming, ongoing, completed, cancelled]
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        isTournament:
          type: boolean
        prize:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Registration:
      type: object
      required: [id, eventId, userId, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        event:
          $ref: "#/components/schemas/Event"
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        teamName:
          type: ["string", "null"]
        teamSize:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    RegistrationInput:
      type: object
      properties:
        teamName:
          type: ["string", "null"]
        teamSize:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
...