# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * NearConnect API - local events discovery and community platform
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  teamSize?: number | null;
  /** @nullable */
  notes?: string | null;
}